add predictive modeling for regression, change output of score_all, fix tree fit parameters
